RDM-2811 - removed unnecessary AuthorisedCaseDataFilter
