+++ src/test/java/uk/gov/hmcts/ccd/domain/service/search/elasticsearch/AuthorisedCaseDetailsSearchOperationTest.java
@@ -1,27 +1,37 @@
 package uk.gov.hmcts.ccd.domain.service.search.elasticsearch;
 
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Map;
 import java.util.Optional;
+import java.util.Set;
 
 import static java.util.Collections.singletonList;
 import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.hasSize;
 import static org.junit.jupiter.api.Assertions.assertAll;
+import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 import static uk.gov.hmcts.ccd.domain.service.common.AccessControlService.CAN_READ;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import org.hamcrest.Matchers;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
+import uk.gov.hmcts.ccd.data.user.UserRepository;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseDetails;
 import uk.gov.hmcts.ccd.domain.model.definition.CaseType;
 import uk.gov.hmcts.ccd.domain.model.search.CaseDetailsSearchResult;
-import uk.gov.hmcts.ccd.domain.service.security.AuthorisedCaseDataFilter;
+import uk.gov.hmcts.ccd.domain.service.common.AccessControlService;
+import uk.gov.hmcts.ccd.domain.service.common.ObjectMapperService;
+import uk.gov.hmcts.ccd.domain.service.common.SecurityClassificationService;
 import uk.gov.hmcts.ccd.domain.service.security.AuthorisedCaseDefinitionDataService;
 
 class AuthorisedCaseDetailsSearchOperationTest {
@@ -34,7 +44,13 @@ class AuthorisedCaseDetailsSearchOperationTest {
     @Mock
     private AuthorisedCaseDefinitionDataService authorisedCaseDefinitionDataService;
     @Mock
-    private AuthorisedCaseDataFilter authorisedCaseDataFilter;
+    private UserRepository userRepository;
+    @Mock
+    private AccessControlService accessControlService;
+    @Mock
+    private SecurityClassificationService classificationService;
+    @Mock
+    private ObjectMapperService objectMapperService;
 
     @InjectMocks
     private AuthorisedCaseDetailsSearchOperation authorisedCaseDetailsSearchOperation;
@@ -55,14 +71,30 @@ class AuthorisedCaseDetailsSearchOperationTest {
         when(authorisedCaseDefinitionDataService.getAuthorisedCaseType(CASE_TYPE, CAN_READ)).thenReturn(Optional.of(caseType));
         when(caseDetailsSearchOperation.execute(CASE_TYPE, QUERY)).thenReturn(searchResult);
 
+        Map<String, JsonNode> unFilteredData = new HashMap<>();
+        caseDetails.setData(unFilteredData);
+        JsonNode jsonNode = mock(JsonNode.class);
+        Set<String> userRoles = new HashSet<>();
+        when(userRepository.getUserRoles()).thenReturn(userRoles);
+        when(objectMapperService.convertObjectToJsonNode(unFilteredData)).thenReturn(jsonNode);
+        CaseType caseType = new CaseType();
+        when(accessControlService.filterCaseFieldsByAccess(jsonNode, caseType.getCaseFields(), userRoles, CAN_READ)).thenReturn(jsonNode);
+        Map<String, JsonNode> filteredData = new HashMap<>();
+        when(objectMapperService.convertJsonNodeToMap(jsonNode)).thenReturn(filteredData);
+
         CaseDetailsSearchResult result = authorisedCaseDetailsSearchOperation.execute(CASE_TYPE, QUERY);
 
         assertAll(
             () -> assertThat(result, is(searchResult)),
+            () -> assertThat(caseDetails.getData(), Matchers.is(filteredData)),
             () -> assertThat(result.getTotal(), is(1L)),
             () -> verify(authorisedCaseDefinitionDataService).getAuthorisedCaseType(CASE_TYPE, CAN_READ),
             () -> verify(caseDetailsSearchOperation).execute(CASE_TYPE, QUERY),
-            () -> verify(authorisedCaseDataFilter).filterFields(caseType, caseDetails)
+            () -> verify(userRepository).getUserRoles(),
+            () -> verify(objectMapperService).convertObjectToJsonNode(unFilteredData),
+            () -> verify(accessControlService).filterCaseFieldsByAccess(jsonNode, caseType.getCaseFields(), userRoles, CAN_READ),
+            () -> verify(objectMapperService).convertJsonNodeToMap(jsonNode),
+            () -> verify(classificationService).applyClassification(caseDetails)
         );
     }
 
@@ -78,7 +110,10 @@ class AuthorisedCaseDetailsSearchOperationTest {
             () -> assertThat(result.getTotal(), is(0L)),
             () -> verify(authorisedCaseDefinitionDataService).getAuthorisedCaseType(CASE_TYPE, CAN_READ),
             () -> verifyZeroInteractions(caseDetailsSearchOperation),
-            () -> verifyZeroInteractions(authorisedCaseDataFilter)
+            () -> verifyZeroInteractions(objectMapperService),
+            () -> verifyZeroInteractions(accessControlService),
+            () -> verifyZeroInteractions(userRepository),
+            () -> verifyZeroInteractions(classificationService)
         );
     }
 }
